pyproject.toml
[tool.poetry]
name = "deep-medical-assignment"
version = "0.1.0"
description = ""
authors = ["Samson Ilemobayo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlmodel = "^0.0.22"
pytest = "^8.3.3"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pydantic = "^2.0"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
